backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
celery==5.3.6
redis==5.0.1
chromadb==0.4.22
openai==1.12.0
python-gmail==1.0.8
langchain==0.1.5
pydantic==2.6.1
python-dotenv==1.0.1
gunicorn==21.2.0